pyproject.toml
[tool.poetry]
name = "codexaudio-service"
version = "0.1.0"
description = "FastAPI service for transcribing MP3 files using OpenAI Whisper, and creating meeting summary MD files using Gemmini-3-4b-it."
authors = ["Viszy A <viszya@viszya.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<=3.13.5"
fastapi = "^0.110.0"
pydub = "*"
pyaudio = "*"
python-dotenv = "*"
python-multipart = "*"
pillow = "*"
torch = "^2.3.0"
transformers = "*"
tqdm = "*"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
openai-whisper = "*"
llvmlite = ">=0.44.0"
numba = ">=0.61.0"
numpy = "<=2.2"

[tool.poetry.group.dev.dependencies]
pytest = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"